==> 2026_Miao_SocialInteractions_ToM/analysis_behavior/preprocessing/6_findInvalidTimePoints.ipynb <==
## Impute number of valid time points

Zizhuang Miao

A problem of continuous ratings is that participants most of the time could not provide valid ratings around the start of each trial; using data from a very small number of participants could lead to biased or wrong estimations of the group judgments. For that, I set the time points with more than 50% of missing data as invalid and mark them by a new indicator variable.

import pandas as pd
import numpy as np
from os.path import join

dataDir = 'C:\\'
threshold = 0.5    # change as needed

##### social interactions

# mean
for n in range(1,9):
    oriData = pd.read_csv(join(dataDir, 'combined_social_consensus', f'narrative{n}_mean.csv'))
    valid_threshold = []
    for s in range(1,10):
        trialData = oriData.loc[oriData.situation==s]
        valid_threshold.append(round(trialData.iloc[0, 3] * threshold))
    for i in range(len(oriData)):
        oriData.loc[i, 'included'] = 0 if oriData.loc[i, 'validPar_time'] < valid_threshold[oriData.loc[i, 'situation']-1] else 1
    oriData.to_csv(join(dataDir, 'combined_social_consensus', f'narrative{n}_mean.csv'), index=False)

# median
for n in range(1,9):
    oriData = pd.read_csv(join(dataDir, 'combined_social_consensus', f'narrative{n}_median.csv'))
    valid_threshold = []
    for s in range(1,10):
        trialData = oriData.loc[oriData.situation==s]
        valid_threshold.append(round(trialData.iloc[0, 3] * threshold))
    for i in range(len(oriData)):
        oriData.loc[i, 'included'] = 0 if oriData.loc[i, 'validPar_time'] < valid_threshold[oriData.loc[i, 'situation']-1] else 1
    oriData.to_csv(join(dataDir, 'combined_social_consensus', f'narrative{n}_median.csv'), index=False)


##### theory of mind

# mean
for n in range(1,9):
    oriData = pd.read_csv(join(dataDir, 'combined_tom_consensus', f'narrative{n}_mean.csv'))
    valid_threshold = []
    for s in range(1,10):
        trialData = oriData.loc[oriData.situation==s]
        valid_threshold.append(round(trialData.iloc[0, 3] * threshold))
    for i in range(len(oriData)):
        oriData.loc[i, 'included'] = 0 if oriData.loc[i, 'validPar_time'] < valid_threshold[oriData.loc[i, 'situation']-1] else 1
    oriData.to_csv(join(dataDir, 'combined_tom_consensus', f'narrative{n}_mean.csv'), index=False)


# median
for n in range(1,9):
    oriData = pd.read_csv(join(dataDir, 'combined_tom_consensus', f'narrative{n}_median.csv'))
    valid_threshold = []
    for s in range(1,10):
        trialData = oriData.loc[oriData.situation==s]
        valid_threshold.append(round(trialData.iloc[0, 3] * threshold))
    for i in range(len(oriData)):
        oriData.loc[i, 'included'] = 0 if oriData.loc[i, 'validPar_time'] < valid_threshold[oriData.loc[i, 'situation']-1] else 1
    oriData.to_csv(join(dataDir, 'combined_tom_consensus', f'narrative{n}_median.csv'), index=False)
